GPU: apply CVE-2024-0671 patch [1/1]
PD#SWPL-176533

Problem:
The problem was that a small memory alias of a large memory allocation
could satisfy both conditions if its GPU VA happens to be aligned to
a large memory page. This would lead to the creation of a 2 MB mapping
even if the user only requested a smaller region, e.g. 4 kB.

Solution:
The driver now adds a third condition to the creation of a MMU ATE
at level 2: the total number of MMU page table entries to map must be
512, i.e. the user has requested a whole 2 MB region intentionally.

Verify:
bifrost r44p0 devices

Change-Id: Ie62a6e0b64d592f0c819525f39d23b20f21f2737
Signed-off-by: yang.zhou <yang.zhou@amlogic.com>