GPU: apply CVE-2023-6363 patch [1/1]
PD#SWPL-183572

Problem:
This patchset defers the registration of a KCPU queue to the KCPU
context until after both the queue and the metadata are completely set
up. This fixes a use-after free, where this dangling pointer is left in
the context. To simplify the control flow, the initialization of the
metadata object has been moved to a separate function.

Solution:
This patchset defers the registration of a KCPU queue to the KCPU context
until after both the queue and the metadata are completely set up
sync from https://scgit.amlogic.com/#/c/457391/

Verify:
valhall r44p1 devices

Change-Id: I8638b81ee4c35b6ad9e49f2c51a57166db359404
Signed-off-by: yajing.li <yajing.li@amlogic.com>